Shutdown cluster in migrate & initialize
This should be done to release resources, which is relevant when used
in a running application (e.g. Play).

Another great thing to accept an existing session in CassandraMigrator,
so that the application using it could just pass the session its
creating anyways. I didn't do this because migrate/initialize right
now accept a DataStore which would then be obsolete, at least it's not
straight forward to allow both - probably there would be 2 constructors:
1 accepting a DataStore and another one accepting a session.