Add cross publishing for scala 2.10 and 2.11.
argot and scalatest had to be updated for this. For argot there's not
yet a publicly available build for scala 2.11 (see bmc/argot#11), therefore
for now I've published it to my bintray repo. As soon as argot is
available for 2.11 in maven central, the argot version can be updated
and the resolver for bintray can be removed.